# backend/functions/openapi.yaml
...
openapi: 3.0.3
info:
  title: Sparkling REST API
  version: 1.0.0
  description: >-
    Sparkling platform backend (SRS v1.0). All JSON, snake_case, ISO-8601 UTC timestamps, money in integer ZAR cents. Mutating routes accept `Idempotency-Key` (or `client_op_id` in the body); replays
    return the cached first response. Every response carries `X-Correlation-Id`. Clients send `X-Client-App` and `X-Client-Version`.
  contact:
    name: Sparkling engineering
    email: dev@sparkling.co.za
  license:
    name: Proprietary
    url: https://sparkling.co.za/legal
servers:
  - url: https://europe-west1-sparkling-4e89d.cloudfunctions.net/api/v1
    description: dev (Cloud Functions 2nd gen)
  - url: http://127.0.0.1:5001/sparkling-4e89d/europe-west1/api/v1
    description: Firebase emulator
...
security:
  - bearerAuth: []
paths:
...
  /memberships/plans:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MembershipPlan'
                  current_plan_code:
                    type: string
                    nullable: true
                  current_status:
                    type: string
                    enum:
                      - pending
                      - active
                      - past_due
                      - cancelled
                      - expired
                    nullable: true
                required:
                  - data
                  - current_plan_code
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Memberships
      operationId: listMembershipPlans
      summary: Active plans with groups → entitlements → services, plus the caller’s current plan
  /memberships/me:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipSummary'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Memberships
      operationId: getMyMembership
      summary: Live membership, plan, selections, allowances (remaining per entitlement in its current period), open/last invoices
  /memberships:
    post:
      responses:
        '200':
          description: Replay of the same client_op_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipSubscribed'
        '201':
          description: Created (status pending)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipSubscribed'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Memberships
      operationId: subscribeMembership
      summary: >-
        Subscribe: pending membership + first invoice + sandbox payment intent. Paying it (sandbox-confirm / provider webhook) activates the membership (period = now → +1 month, tier = plan,
        membership_activated). 409 conflict when a live membership exists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipSubscribe'
  /memberships/me/invoices/{id}/pay:
    post:
      responses:
        '200':
          description: Open intent replayed
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoice:
                    $ref: '#/components/schemas/MembershipInvoice'
                  payment:
                    $ref: '#/components/schemas/MembershipPayment'
                  duplicate:
                    type: boolean
        '201':
          description: Intent created
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoice:
                    $ref: '#/components/schemas/MembershipInvoice'
                  payment:
                    $ref: '#/components/schemas/MembershipPayment'
                  duplicate:
                    type: boolean
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Memberships
      operationId: payMembershipInvoice
      summary: Sandbox payment intent for a pending renewal/upgrade invoice (a still-open intent is replayed); success rolls the period (membership_renewed)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
  /memberships/me/selections:
    put:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipSummary'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Memberships
      operationId: changeMembershipSelections
      summary: Change the chosen option per choose_one group; 409 once any allowance was used this period
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                selections:
                  type: object
                  additionalProperties:
                    type: string
              required:
                - selections
  /memberships/me/change-plan:
    post:
      responses:
        '200':
          description: Downgrade scheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipChangePlan'
        '201':
          description: Upgrade invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipChangePlan'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Memberships
      operationId: changeMembershipPlan
      summary: 'Upgrade (higher fee): full-fee invoice + intent now, plan and a fresh period start on payment. Downgrade: next_plan_id applied at renewal. 409 when already on the plan or not active'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                plan_code:
                  type: string
                selections:
                  type: object
                  additionalProperties:
                    type: string
              required:
                - plan_code
  /memberships/me/cancel:
    post:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  membership:
                    $ref: '#/components/schemas/Membership'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Memberships
      operationId: cancelMembership
      summary: Cancel at period end (default; benefits continue until then) or immediately (cancelled, tier → silver, pending invoices void). membership_cancelled
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                at_period_end:
                  type: boolean
                  default: true
                reason:
                  type: string
                  nullable: true
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
  responses:
    ValidationError:
      description: validation_error (400)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthenticated:
      description: unauthenticated (401)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: forbidden (403)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: not_found (404)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: conflict / invalid_transition (409)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Gone:
      description: gone (410) — expired quote link or quotation past valid_until
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RateLimited:
      description: rate_limited (429)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Internal:
      description: internal (500) — never includes stack traces
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - unauthenticated
                - forbidden
                - not_found
                - validation_error
                - conflict
                - invalid_transition
                - invalid_otp
                - gone
                - rate_limited
                - internal
            message:
              type: string
            details:
              type: array
              items: {}
            correlation_id:
              type: string
          required:
            - code
            - message
            - correlation_id
      required:
        - error
...
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        booking_id:
          type: string
          format: uuid
          nullable: true
        quotation_id:
          type: string
          format: uuid
          nullable: true
        customer_id:
          type: string
        provider:
          type: string
        provider_ref:
          type: string
          nullable: true
        method_id:
          type: string
          format: uuid
          nullable: true
        amount_cents:
          type: integer
        currency:
          type: string
        status:
          type: string
          enum:
            - initiated
            - pending
            - successful
            - failed
            - cancelled
            - refunded
        receipt_no:
          type: string
          example: RCP-70001
          nullable: true
        idempotency_key:
          type: string
        failure_reason:
          type: string
          nullable: true
        verified_at:
          type: string
          format: date-time
          nullable: true
        method:
          type: string
          enum:
            - cash
            - card_terminal
            - card
            - eft
          nullable: true
        recorded_by:
          type: string
          description: Staff uid for POS attestations
          nullable: true
        membership_invoice_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    MembershipEntitlement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: G1
        label:
          type: string
          example: 4 × Sparkling Wash
        quantity:
          type: integer
        period:
          type: string
          enum:
            - month
            - year
        sort_order:
          type: integer
        services:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              code:
                type: string
                example: SPARKLING_WASH
              name:
                type: string
              is_primary:
                type: boolean
      required:
        - id
        - code
        - label
        - quantity
        - period
        - services
    MembershipGroup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: washes
        name:
          type: string
        selection:
          type: string
          enum:
            - choose_one
            - all
        sort_order:
          type: integer
        entitlements:
          type: array
          items:
            $ref: '#/components/schemas/MembershipEntitlement'
      required:
        - id
        - code
        - name
        - selection
        - entitlements
    MembershipPlan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: gold
        tier:
          type: string
          enum:
            - silver
            - gold
            - platinum
            - black
        name:
          type: string
        tagline:
          type: string
          nullable: true
        monthly_fee_cents:
          type: integer
          example: 29500
        discount_pct:
          type: number
          example: 10
        discount_scope:
          type: string
          enum:
            - plan_services
            - other_services
            - all_services
            - none
        discount_note:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        sort_order:
          type: integer
        is_active:
          type: boolean
        groups:
          type: array
          items:
            $ref: '#/components/schemas/MembershipGroup'
      required:
        - id
        - code
        - tier
        - name
        - monthly_fee_cents
        - discount_pct
        - discount_scope
        - groups
...
    Membership:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ref:
          type: string
          example: MEM-2026-0001
        customer_id:
          type: string
        plan_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - pending
            - active
            - past_due
            - cancelled
            - expired
        started_at:
          type: string
          format: date-time
          nullable: true
        current_period_start:
          type: string
          format: date-time
          nullable: true
        current_period_end:
          type: string
          format: date-time
          nullable: true
        cancel_at_period_end:
          type: boolean
        cancelled_at:
          type: string
          format: date-time
          nullable: true
        ended_at:
          type: string
          format: date-time
          nullable: true
        next_plan_id:
          type: string
          format: uuid
          nullable: true
        payment_method:
          type: string
          enum:
            - card
            - cash
            - eft
            - sandbox
        client_op_id:
          type: string
          nullable: true
        created_by:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - ref
        - customer_id
        - plan_id
        - status
        - cancel_at_period_end
        - payment_method
    MembershipInvoice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ref:
          type: string
          example: MINV-2026-0101
        membership_id:
          type: string
          format: uuid
        customer_id:
          type: string
        period_start:
          type: string
          format: date-time
        period_end:
          type: string
          format: date-time
        amount_cents:
          type: integer
        status:
          type: string
          enum:
            - pending
            - paid
            - failed
            - void
        due_at:
          type: string
          format: date-time
        paid_at:
          type: string
          format: date-time
          nullable: true
        payment_id:
          type: string
          format: uuid
          nullable: true
        idempotency_key:
          type: string
          example: renewal:<membership_id>:2026-10-10
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - ref
        - membership_id
        - period_start
        - period_end
        - amount_cents
        - status
        - due_at
    Allowance:
      type: object
      properties:
        entitlement_id:
          type: string
          format: uuid
        entitlement_code:
          type: string
          example: G1
        group_code:
          type: string
          example: washes
        label:
          type: string
        quantity:
          type: integer
        used:
          type: integer
        remaining:
          type: integer
        period:
          type: string
          enum:
            - month
            - year
        period_start:
          type: string
          format: date-time
        period_end:
          type: string
          format: date-time
      required:
        - entitlement_id
        - entitlement_code
        - group_code
        - label
        - quantity
        - used
        - remaining
        - period
        - period_start
        - period_end
...
    MembershipSummary:
      type: object
      properties:
        membership:
          type: object
          allOf:
            - $ref: '#/components/schemas/Membership'
          nullable: true
        plan:
          type: object
          allOf:
            - $ref: '#/components/schemas/MembershipPlan'
          nullable: true
        next_plan:
          type: object
          properties:
            id:
              type: string
              format: uuid
            code:
              type: string
            name:
              type: string
            monthly_fee_cents:
              type: integer
          nullable: true
        selections:
          type: object
          additionalProperties:
            type: string
          description: '{ [group_code]: entitlement_code }'
        allowances:
          type: array
          items:
            $ref: '#/components/schemas/Allowance'
        open_invoice:
          type: object
          allOf:
            - $ref: '#/components/schemas/MembershipInvoice'
          nullable: true
        invoices:
          type: array
          items:
            $ref: '#/components/schemas/MembershipInvoice'
          description: Last 12, newest first
        next_renewal_at:
          type: string
          format: date-time
          nullable: true
        benefits_summary:
          type: string
          example: 3 of 4 Sparkling Washes left · 10% discount on any other Sparkling service
      required:
        - membership
        - plan
        - selections
        - allowances
        - open_invoice
        - invoices
        - next_renewal_at
        - benefits_summary
    MembershipPayment:
      allOf:
        - $ref: '#/components/schemas/Payment'
        - type: object
          properties:
            client_secret:
              type: string
              description: Sandbox intent secret (null on replay)
              nullable: true
    MembershipSubscribe:
      type: object
      properties:
        plan_code:
          type: string
          example: gold
        selections:
          type: object
          additionalProperties:
            type: string
          example:
            washes: G1
        payment_method:
          type: string
          enum:
            - card
        client_op_id:
          type: string
      required:
        - plan_code
        - client_op_id
    MembershipSubscribed:
      type: object
      properties:
        membership:
          $ref: '#/components/schemas/Membership'
        invoice:
          $ref: '#/components/schemas/MembershipInvoice'
        payment:
          type: object
          allOf:
            - $ref: '#/components/schemas/MembershipPayment'
          nullable: true
        duplicate:
          type: boolean
      required:
        - membership
        - invoice
        - payment
        - duplicate
...
    MembershipChangePlan:
      type: object
      properties:
        membership:
          $ref: '#/components/schemas/Membership'
        change:
          type: string
          enum:
            - upgrade
            - downgrade
        invoice:
          type: object
          allOf:
            - $ref: '#/components/schemas/MembershipInvoice'
          nullable: true
        payment:
          type: object
          allOf:
            - $ref: '#/components/schemas/MembershipPayment'
          nullable: true
        applies_at:
          type: string
          format: date-time
          nullable: true
      required:
        - membership
        - change
        - invoice
        - payment
        - applies_at
...